requirements.txt
fabric
jinja2
django-fabdeploy-plus
